+++ index.html
@@ -29,9 +29,6 @@
             <button class="controls_color jsColor" style="background-color: #5856d6"></button>
         </div>
     </div>
-    <!-- <script src="app.js"></script> -->
-    <script src="Canvas.js"></script>
-    <script src="Palette.js"></script>
-    <script src="Brush.js"></script>
+    <script src="./src/app.js" type="module"></script>
 </body>
 </html>
